Attendees moved to subcollections, multiple simultaneous sheets
attendees are now subcollections of the Sheets doc, instead of arrays, so that we can store more information, such as checkin and checkout information. 
Simultaneous logging of multiple sheets now works. 
Finding the Correct Sheet and Creating a sheet if the correct one doesn't exist now both happen in the same function.